+++ tests/core/test_core.py
@@ -33,7 +33,7 @@ def fixture_candles_31st(candles):
 def test_data(fixture_candles_30, expected_ema):
     test = EMA(candles=fixture_candles_30, source="close")
     test.calculate()
-    assert pytest.approx(test.readings()) == expected_ema[0:30]
+    assert pytest.approx(test.series()) == expected_ema[0:30]
 
 
 def test_data_append(fixture_candles_30, fixture_candles_31st, expected_ema):
@@ -42,4 +42,4 @@ def test_data_append(fixture_candles_30, fixture_candles_31st, expected_ema):
     test.append(fixture_candles_31st)
     test.calculate()
 
-    assert pytest.approx(test.readings()) == expected_ema[0:31]
+    assert pytest.approx(test.series()) == expected_ema[0:31]