+++ tests/core/test_hexital_timeframe.py
@@ -7,24 +7,24 @@ from hexital.indicators import EMA, OBV, SMA
 def test_hextial_multi_timeframes(candles, expected_ema, expected_sma_t10):
     strategy = Hexital("Test Strategy", candles, [EMA(), SMA(timeframe="t10")])
     strategy.calculate()
-    assert pytest.approx(strategy.reading_as_list("EMA_10")) == expected_ema
-    assert pytest.approx(strategy.reading_as_list("SMA_10_T10")) == expected_sma_t10
+    assert pytest.approx(strategy.series("EMA_10")) == expected_ema
+    assert pytest.approx(strategy.series("SMA_10_T10")) == expected_sma_t10
 
 
 def test_hextial_multi_timeframes_append(candles, expected_ema, expected_sma_t10):
     strategy = Hexital("Test Strategy", candles[:251], [EMA(), SMA(timeframe="t10")])
     strategy.calculate()
 
-    assert pytest.approx(strategy.reading_as_list("EMA_10")) == expected_ema[:251]
+    assert pytest.approx(strategy.series("EMA_10")) == expected_ema[:251]
     assert (
-        pytest.approx(strategy.reading_as_list("SMA_10_T10"), 1.5e-1)
+        pytest.approx(strategy.series("SMA_10_T10"), 1.5e-1)
         == expected_sma_t10[:26]
     )
 
     strategy.append(candles[-249:])
 
-    assert pytest.approx(strategy.reading_as_list("EMA_10")) == expected_ema
-    assert pytest.approx(strategy.reading_as_list("SMA_10_T10")) == expected_sma_t10
+    assert pytest.approx(strategy.series("EMA_10")) == expected_ema
+    assert pytest.approx(strategy.series("SMA_10_T10")) == expected_sma_t10
 
 
 def test_hextial_multi_timeframes_shared_candles(
@@ -43,7 +43,7 @@ def test_hextial_multi_timeframes_shared_candles(
         pytest.fail("Expected timeframe candles for t10 strategy")
 
     assert len(strategy.get_candles()) == 2
-    assert pytest.approx(strategy.reading_as_list("EMA_10")) == expected_ema
+    assert pytest.approx(strategy.series("EMA_10")) == expected_ema
     assert (
         t10_candles[-1].indicators.get("SMA_10_T10") == expected_sma_t10[-1]
         and t10_candles[-1].indicators.get("OBV_T10") == expected_obv_t10[-1]
@@ -82,9 +82,9 @@ def test_hextial_multi_timeframes_lifespan(candles, expected_ema, expected_sma_t
     for candle in candles:
         strategy.append(candle)
 
-    assert pytest.approx(strategy.reading_as_list("EMA_10")) == expected_ema[-61:]
+    assert pytest.approx(strategy.series("EMA_10")) == expected_ema[-61:]
     assert (
-        pytest.approx(strategy.reading_as_list("SMA_10_T5"), 1.5e-02)
+        pytest.approx(strategy.series("SMA_10_T5"), 1.5e-02)
         == expected_sma_t5[-13:]
     )